requirements.txt
asgiref==3.10.0
Django==5.2.7
django-autoslug==1.9.9
pillow==11.3.0
sqlparse==0.5.3
